Remove annotations in mapviewcontroller

+++ On the Map/StudentInfo.swift
@@ -12,15 +12,15 @@ struct StudentInfo {
     var firstName: String
     var lastName: String
     var linkUrl: String
-    var latitude: Float
-    var longitude: Float
+    var latitude: Double
+    var longitude: Double
 
     init(dictionary: [String: AnyObject]) {
         firstName = dictionary["firstName"] as! String
         lastName = dictionary["lastName"] as! String
         linkUrl = dictionary["linkUrl"] as! String
-        latitude = dictionary["latitude"] as! Float
-        longitude = dictionary["longitude"] as! Float
+        latitude = dictionary["latitude"] as! Double
+        longitude = dictionary["longitude"] as! Double
     }
     
     func fullName() -> String {
